refactor: access token 검증 로직을 utils로 분리
- access token 검증 로직을 utils로 분리
- 에러 응답 타입을 types.ts에 추가
