+++ src/mocks/types.ts
@@ -1,0 +1,3 @@
+export type ErrorResponse = {
+  message: string
+}
